Remove usages of static_cast<enum> when outputing to stream.
Bug: 1318236
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4673703
Auto-Submit: Xiaoling Bao <xiaolingbao@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Xiaoling Bao <xiaolingbao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1167652}

+++ chrome/updater/app/app_uninstall.cc
@@ -169,8 +169,7 @@ void AppUninstall::UninstallAll(int reason) {
           [](base::OnceCallback<void(int)> shutdown, UpdaterScope scope,
              update_client::Error uninstall_ping_error) {
             VLOG_IF(1, uninstall_ping_error != update_client::Error::NONE)
-                << "Uninstall ping failed: "
-                << static_cast<int>(uninstall_ping_error);
+                << "Uninstall ping failed: " << uninstall_ping_error;
             base::ThreadPool::PostTaskAndReplyWithResult(
                 FROM_HERE, {base::MayBlock()},
                 base::BindOnce(
