+++ chrome/updater/remove_uninstalled_apps_task.cc
@@ -83,7 +83,7 @@ std::vector<PingInfo> GetAppIDsToRemove(
 void UninstallPingSent(base::RepeatingClosure callback,
                        update_client::Error error) {
   if (error != update_client::Error::NONE)
-    VLOG(0) << __func__ << ": Error: " << static_cast<int>(error);
+    VLOG(0) << __func__ << ": Error: " << error;
   callback.Run();
 }
